move decomposition methods to a newer folder
