+++ gwlearn/decomposition/__init__.py
@@ -1,0 +1,4 @@
+from ._base import BaseDecomposition
+from .pca import GWPCA
+
+__all__ = ["BaseDecomposition", "GWPCA"]
